refactor code to fully hide zlib_filter definition
